+++ src/output/variables/precipitation.jl
@@ -19,11 +19,11 @@ path(::ConvectivePrecipitationOutput, simulation) =
 # at finalize step postprocess the convective precipitation to get the rate
 finalize!(output::NetCDFOutput, variable::ConvectivePrecipitationOutput, args...) = output!(output, variable.rate, variable)
 
-abstract type AbstractRainRateOutputVariable <: AbstractOutputVariable end
+abstract type AbstractRateOutputVariable <: AbstractOutputVariable end
 
 """Defines netCDF output for a specific variables, see [`VorticityOutput`](@ref) for details.
 Fields are: $(TYPEDFIELDS)"""
-@kwdef mutable struct ConvectivePrecipitationRateOutput{F} <: AbstractRainRateOutputVariable
+@kwdef mutable struct ConvectivePrecipitationRateOutput{F} <: AbstractRateOutputVariable
     name::String = "precip_conv_rate"
     unit::String = "mm/hr"
     long_name::String = "convective precipitation rate"
@@ -35,9 +35,12 @@ Fields are: $(TYPEDFIELDS)"""
     transform::F = (x) -> 1000x     # [m] to [mm]
 end
 
+"""$TYPEDSIGNATURES
+Post-process the netCDF `output` file to convert accumulated precipitation/snow to
+rates."""
 function output!(
     output::NetCDFOutput,
-    variable::AbstractRainRateOutputVariable,
+    variable::AbstractRateOutputVariable,
     acc_variable::AbstractOutputVariable,
 )
     # use .var to prevent Union{Missing, Float32} that NCDatasets uses
@@ -80,10 +83,80 @@ finalize!(output::NetCDFOutput, variable::LargeScalePrecipitationOutput, args...
 
 """Defines netCDF output for a specific variables, see [`VorticityOutput`](@ref) for details.
 Fields are: $(TYPEDFIELDS)"""
-@kwdef mutable struct LargeScalePrecipitationRateOutput{F} <: AbstractRainRateOutputVariable
+@kwdef mutable struct LargeScalePrecipitationRateOutput{F} <: AbstractRateOutputVariable
     name::String = "precip_cond_rate"
     unit::String = "mm/hr"
     long_name::String = "large-scale precipitation rate"
+    dims_xyzt::NTuple{4, Bool} = (true, true, false, true)
+    missing_value::Float64 = NaN
+    compression_level::Int = 3
+    shuffle::Bool = true
+    keepbits::Int = 7
+    transform::F = (x) -> 1000x     # [m] to [mm]
+end
+
+"""Defines netCDF output for a specific variables, see [`VorticityOutput`](@ref) for details.
+Fields are: $(TYPEDFIELDS)"""
+@kwdef mutable struct ConvectiveSnowOutput{F, R} <: AbstractOutputVariable
+    name::String = "snow_conv"
+    unit::String = "mm"
+    long_name::String = "accumulated convective snow"
+    dims_xyzt::NTuple{4, Bool} = (true, true, false, true)
+    missing_value::Float64 = NaN
+    compression_level::Int = 3
+    shuffle::Bool = true
+    keepbits::Int = 20
+    transform::F = (x) -> 1000x                     # [m] to [mm]
+    rate::R = ConvectiveSnowRateOutput()   # include here to be called at finalize!
+end
+
+path(::ConvectiveSnowOutput, simulation) =
+    simulation.diagnostic_variables.physics.snow_convection
+
+# at finalize step postprocess the convective snow to get the rate
+finalize!(output::NetCDFOutput, variable::ConvectiveSnowOutput, args...) = output!(output, variable.rate, variable)
+
+"""Defines netCDF output for a specific variables, see [`VorticityOutput`](@ref) for details.
+Fields are: $(TYPEDFIELDS)"""
+@kwdef mutable struct ConvectiveSnowRateOutput{F} <: AbstractRateOutputVariable
+    name::String = "snow_conv_rate"
+    unit::String = "mm/hr"
+    long_name::String = "convective snow rate"
+    dims_xyzt::NTuple{4, Bool} = (true, true, false, true)
+    missing_value::Float64 = NaN
+    compression_level::Int = 3
+    shuffle::Bool = true
+    keepbits::Int = 7
+    transform::F = (x) -> 1000x     # [m] to [mm]
+end
+
+"""Defines netCDF output for a specific variables, see [`VorticityOutput`](@ref) for details.
+Fields are: $(TYPEDFIELDS)"""
+@kwdef mutable struct LargeScaleSnowOutput{F, R} <: AbstractOutputVariable
+    name::String = "snow_cond"
+    unit::String = "mm"
+    long_name::String = "accumulated large-scale snow"
+    dims_xyzt::NTuple{4, Bool} = (true, true, false, true)
+    missing_value::Float64 = NaN
+    compression_level::Int = 3
+    shuffle::Bool = true
+    keepbits::Int = 20
+    transform::F = (x) -> 1000x                     # [m] to [mm]
+    rate::R = LargeScaleSnowRateOutput()   # include here to be called at finalize!
+end
+
+path(::LargeScaleSnowOutput, simulation) =
+    simulation.diagnostic_variables.physics.snow_large_scale
+
+# at finalize step postprocess the convective snow precipitation to get the rate
+finalize!(output::NetCDFOutput, variable::LargeScaleSnowOutput, args...) = output!(output, variable.rate, variable)
+
+"""Defines netCDF output for a specific variables, see [`VorticityOutput`](@ref) for details.
+Fields are: $(TYPEDFIELDS)"""
+@kwdef mutable struct LargeScaleSnowRateOutput{F} <: AbstractRateOutputVariable
+    name::String = "snow_cond_rate"
+    unit::String = "mm/hr"
+    long_name::String = "large-scale snow rate"
     dims_xyzt::NTuple{4, Bool} = (true, true, false, true)
     missing_value::Float64 = NaN
     compression_level::Int = 3
@@ -112,5 +185,7 @@ path(::CloudTopOutput, simulation) =
 PrecipitationOutput() = (
     ConvectivePrecipitationOutput(),
     LargeScalePrecipitationOutput(),
+    ConvectiveSnowOutput(),
+    LargeScaleSnowOutput(),
     CloudTopOutput(),
 )