+++ src/physics/column_variables.jl
@@ -150,10 +150,19 @@ function write_column_tendencies!(
     diagn.physics.precip_rate_large_scale[ij] = column.precip_rate_large_scale
     diagn.physics.precip_rate_convection[ij] = column.precip_rate_convection
 
-    # total precipitation rate [kg/m²/s]
+    # accumulate snow [m]
+    diagn.physics.snow_large_scale[ij] += column.snow_large_scale
+    diagn.physics.snow_convection[ij] += column.snow_convection
+
+    # precipitation rate [m/s], instantaneous (i.e. overwrite, do not accumulate)
+    diagn.physics.snow_rate_large_scale[ij] = column.snow_rate_large_scale
+    diagn.physics.snow_rate_convection[ij] = column.snow_rate_convection
+    
+    # total precipitation rate (rain+snow) [kg/m²/s]
     ρ = atmosphere.water_density
     diagn.physics.total_precipitation_rate[ij] =
-        (column.precip_rate_large_scale + column.precip_rate_convection) * ρ
+        (column.precip_rate_large_scale + column.precip_rate_convection +
+         column.snow_rate_large_scale + column.snow_rate_convection) * ρ
 
     # Cloud top in height [m] from geopotential height divided by gravity, 0 for no clouds
     diagn.physics.cloud_top[ij] = column.cloud_top == nlayers+1 ? 0 : column.geopot[column.cloud_top]
@@ -222,12 +231,18 @@ function reset_column!(column::ColumnVariables{NF}) where NF
     column.flux_temp_upward .= 0
     column.flux_temp_downward .= 0
 
-    # Convection and precipitation
+    # Convection and large-scale precipitation
     column.cloud_top = column.nlayers+1
     column.precip_convection = 0            # set back to zero to accumulate in the vertical
     column.precip_large_scale = 0
     column.precip_rate_convection = 0       # instantaneously overwritten, but convection may escape early
     column.precip_rate_large_scale = 0
+    
+    # same for snow
+    column.snow_convection = 0              # set back to zero to accumulate in the vertical
+    column.snow_large_scale = 0
+    column.snow_rate_convection = 0         # instantaneously overwritten, but convection may escape early
+    column.snow_rate_large_scale = 0
 
     # radiation
     column.outgoing_longwave_radiation = 0