+++ src/physics/convection.jl
@@ -37,15 +37,21 @@ export SimplifiedBettsMiller
 https://doi.org/10.1175/JAS3935.1. This implements the qref-formulation
 in their paper. Fields and options are
 $(TYPEDFIELDS)"""
-@kwdef struct SimplifiedBettsMiller{NF, SP} <: AbstractConvection
-    "number of vertical layers"
-    nlayers::Int
-
+@kwdef mutable struct SimplifiedBettsMiller{NF, SP} <: AbstractConvection
     "[OPTION] Relaxation time for profile adjustment"
     time_scale::Second = Hour(4)
 
-    "[OPTION] Relative humidity for reference profile"
+    "[OPTION] Relative humidity for reference profile [1]"
     relative_humidity::NF = 0.7
+
+    "[OPTION] Convert rain fall to snow below freezing?"
+    snow::Bool = true
+
+    "[OPTION] Freezing temperature below which to convert rain into snow [K]"
+    freezing_threshold::NF = 263
+
+    "[OPTION] Melting temperature for snow fall [K]"
+    melting_threshold::NF = 278
 
     "[OPTION] Surface perturbation of temp, humid to calculate the moist pseudo adiabat"
     surface_temp_humid::SP  # don't specify default here but in generator below
@@ -57,8 +63,7 @@ function SimplifiedBettsMiller(
     surface_temp_humid = NoSurfacePerturbation(),
     kwargs...)
     # type inference happens here as @kwdef only defines no/all types specified
-    return SimplifiedBettsMiller{SG.NF, typeof(surface_temp_humid)}(
-        nlayers=SG.nlayers; surface_temp_humid, kwargs...)
+    return SimplifiedBettsMiller{SG.NF, typeof(surface_temp_humid)}(; surface_temp_humid, kwargs...)
 end
 
 # nothing to initialize but maybe the surface temp/humid functor?
@@ -105,8 +110,10 @@ function convection!(
     Δσ = geometry.σ_levels_thick
     (; geopot, nlayers, temp, temp_virt, humid, temp_tend, humid_tend) = column
     pₛ = column.pres[end]
-    (; Lᵥ, cₚ) = clausius_clapeyron
-    
+    (; Lᵥ, Lᵢ, cₚ) = clausius_clapeyron
+    let_it_snow = SBM.snow              # flag to switch on/off rain -> snow conversion
+    (; freezing_threshold, melting_threshold) = SBM        # threshold temperature below which it snows
+
     # use work arrays for temp_ref_profile, humid_ref_profile
     temp_ref_profile = column.a     # temperature [K] reference profile to adjust to
     humid_ref_profile = column.b    # specific humidity [kg/kg] profile to adjust to
@@ -128,6 +135,7 @@ function convection!(
     local PT::NF = 0        # precipitation due to cooling
     local ΔT::NF = 0        # vertically uniform temperature profile adjustment
     local Qref::NF = 0      # = ∫_pₛ^p_LZB -humid_ref_profile dp
+    local snow_flux::NF = 0.0   # start an empty snow flux at top of atmosphere
 
     # skip constants compared to Frierson 2007, i.e. no /τ, /gravity, *cₚ/Lᵥ
     for k in level_zero_buoyancy:nlayers
@@ -180,6 +188,11 @@ function convection!(
         end
     end
 
+	(; gravity) = planet
+	(; water_density) = atmosphere
+	(; Δt_sec) = time_stepping
+	pₛΔt_gρ = (pₛ * Δt_sec / gravity / water_density) # enfore no precip for shallow conv
+	
     # GET TENDENCIES FROM ADJUSTED PROFILES
     for k in level_zero_buoyancy:nlayers
         temp_tend[k] -= (temp[k] - temp_ref_profile[k]) / SBM.time_scale.value
@@ -187,15 +200,46 @@ function convection!(
         humid_tend[k] -= δq
 
         # convective precipiation, integrate dq\dt [(kg/kg)/s] vertically
-        column.precip_convection += δq * Δσ[k]
-    end
+        precip = max(δq * Δσ[k], zero(δq))      # only integrate excess humidity for precip (no reevaporation)
+        snow = zero(precip)                     # start with zero snow but potentially swap below
 
-    (; gravity) = planet
-    (; water_density) = atmosphere
-    (; Δt_sec) = time_stepping
-    pₛΔt_gρ = (pₛ * Δt_sec / gravity / water_density) * deep_convection # enfore no precip for shallow conv 
-    column.precip_convection *= pₛΔt_gρ                                 # convert to [m] of rain during Δt
+        # decide whether to turn precip into snow
+        precip, snow = let_it_snow && temp[k] < freezing_threshold ? (snow, precip) : (precip, snow)
+        snow_flux += snow # assign snow to the snow_flux local variable
+        # latent heat release for enthalpy conservation
+        L = snow > 0 ? (Lᵥ + Lᵢ) : Lᵥ
+        δT = -L / cₚ * δq
+
+        # I am not sure whether or not these belong here. They are needed in condensation, perhaps not here?
+        humid_tend[k] += δq
+        temp_tend[k] += δT
+		
+        # now test whether any snow has come into this layer as snow_flux and whether the current layer can melt it
+        δT_melt = temp[k] - melting_threshold
+        if snow_flux > 0 && δT_melt > 0 
+			E_avail = cₚ * δT_melt                            # J / kg (of air)
+			melt_depth = (E_avail / Lᵢ) * (Δσ[k] * pₛΔt_gρ)   # [m]
+			melt_amount = min(snow_flux, melt_depth)          # [m]
+			δq_melt = melt_amount / (Δσ[k] * pₛΔt_gρ)         # [kg/kg]
+			snow_flux    -= melt_amount                       # consistent units
+			precip       += melt_amount
+			δT_melt_actual = -(Lᵢ/cₚ) * δq_melt
+			temp_tend[k] += δT_melt_actual
+		end
+        
+        column.precip_convection += precip     # integrate vertically, Formula 25, unit [m]
+        #column.snow_convection   += snow       # No longer do this here
+    end
+ 
+    pₛΔt_gρ = (pₛ * Δt_sec / gravity / water_density) * deep_convection # enfore no precip for shallow conv
+	column.precip_convection *= pₛΔt_gρ                                 # convert to [m] of rain during Δt
     column.precip_rate_convection = column.precip_convection / Δt_sec   # rate: convert to [m/s] of rain
+
+    # same for snow
+    #column.snow_convection *= pₛΔt_gρ                                   # convert to [m] of rain during Δt
+	column.snow_convection   = snow_flux       # do it here, as snow_flux holds the vertical integral [m]
+    column.snow_rate_convection = column.snow_convection / Δt_sec       # rate: convert to [m/s] of rain
+
     column.cloud_top = min(column.cloud_top, level_zero_buoyancy)       # clouds reach to top of convection
     return nothing
 end
@@ -289,9 +333,6 @@ https://doi.org/10.1175/JAS3935.1 but with humidity set to zero.
 Fields and options are
 $(TYPEDFIELDS)"""
 @kwdef struct DryBettsMiller{NF, SP} <: AbstractConvection
-    "number of vertical layers/levels"
-    nlayers::Int
-
     "[OPTION] Relaxation time for profile adjustment"
     time_scale::Second = Hour(4)
 
@@ -305,7 +346,7 @@ function DryBettsMiller(
     surface_temp = NoSurfacePerturbation(),
     kwargs...)
     # infer type here as @kwdef only defines constructors for no/all types specified
-    return DryBettsMiller{SG.NF, typeof(surface_temp)}(nlayers=SG.nlayers; surface_temp, kwargs...)
+    return DryBettsMiller{SG.NF, typeof(surface_temp)}(; surface_temp, kwargs...)
 end
 
 # nothing to initialize but maybe the surface temp functor?